backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
google-generativeai==0.7.2
pydantic==2.8.2
python-dotenv==1.0.1
